requirements.txt
nltk==3.8.1
numpy==1.26.3
pyenchant==3.2.2
strsimpy==0.2.1
